requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==0.15.0
marshmallow==3.20.2
marshmallow-sqlalchemy==0.29.0
psycopg
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
alembic
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
click
typing-extensions
charset-normalizer
idna
urllib3
